requirements.txt
absl-py==0.13.0
APScheduler==3.6.3
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
comtypes==1.1.10
cycler==0.10.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.2.5
pathlib==1.0.1
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pypiwin32==223
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
python-telegram-bot==13.6
pyttsx3==2.90
pytz==2021.1
pywin32==301
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
